+++ src/main/java/com/ijse/lostandfoundsystem/service/ItemService.java
@@ -6,8 +6,8 @@ import java.util.List;
 
 public interface ItemService {
     void reportItem(ItemDTO itemDTO);
-    ItemDTO getSelectedItem(int itemId);
+    ItemDTO getSelectedItem(Integer itemId);
     List<ItemDTO> getAllItems();
-    void updateItem(Long itemId, ItemDTO itemDTO);
-    void deleteItem(Long itemId);
+    void updateItem(Integer itemId, ItemDTO itemDTO);
+    void deleteItem(Integer itemId);
 }
